Refactor CmServerApi with public interface, factory and private impl
