+++ src/main/java/com/cloudera/whirr/cm/CmServerHandler.java
@@ -35,6 +35,7 @@ import org.apache.whirr.service.ClusterActionEvent;
 import org.apache.whirr.service.FirewallManager.Rule;
 
 import com.cloudera.whirr.cm.api.CmServerApi;
+import com.cloudera.whirr.cm.api.CmServerApiFactory;
 import com.cloudera.whirr.cm.api.CmServerApiLog;
 import com.cloudera.whirr.cm.api.CmServerService;
 import com.cloudera.whirr.cm.api.CmServerServiceType;
@@ -184,8 +185,9 @@ public class CmServerHandler extends BaseHandlerCm {
             }
           }
 
-          CmServerApi cmServerApi = new CmServerApi(event.getCluster().getInstanceMatching(role(ROLE)).getPublicIp(),
-              7180, CM_USER, CM_PASSWORD, new CmServerApiLog.CmServerApiLogSysOut());
+          CmServerApi cmServerApi = CmServerApiFactory.getCmServerApi(event.getCluster()
+              .getInstanceMatching(role(ROLE)).getPublicIp(), 7180, CM_USER, CM_PASSWORD,
+              new CmServerApiLog.CmServerApiLogSysOut());
           cmServerApi.initialise(config);
           cmServerApi.provision(BaseHandlerCmCdh.CmServerClusterSingleton.getInstance());
           cmServerApi.configure(BaseHandlerCmCdh.CmServerClusterSingleton.getInstance());
@@ -234,8 +236,8 @@ public class CmServerHandler extends BaseHandlerCm {
           }
 
           logLineItem("Start:");
-          CmServerApi cmServerApi = new CmServerApi(event.getCluster().getInstanceMatching(role(ROLE)).getPublicIp(),
-              7180, CM_USER, CM_PASSWORD, new CmServerApiLog.CmServerApiLogSysOut());
+          CmServerApi cmServerApi = CmServerApiFactory.getCmServerApi(event.getCluster().getInstanceMatching(role(ROLE))
+              .getPublicIp(), 7180, CM_USER, CM_PASSWORD, new CmServerApiLog.CmServerApiLogSysOut());
           cmServerApi.start(BaseHandlerCmCdh.CmServerClusterSingleton.getInstance());
 
         } catch (Exception e) {
